Update readme.md Update folders and namespaces
