+++ src/JSLTSharp.Tests/EmbededFunctions/DefaultStringTransformTests.cs
