+++ src/JSLTSharp.Tests/EmbededFunctions/IsEqualsTests.cs
