+++ src/JSLTSharp.Tests/EmbededFunctions/RootArrayTransformTests.cs
