+++ src/JSLTSharp.Tests/EmbededFunctions/ValueTransformations/ToBooleanTransformTests.cs
