+++ src/JSLTSharp.Tests/EmbededFunctions/ValueTransformations/ToIntegerTransformTests.cs
