+++ src/JSLTSharp.Tests/EmbededFunctions/ValueTransformations/ToUpperTransformTests.cs
