+++ src/JSLTSharp/JsonTransforms/EmbededFunctions/ConditionalOperations/NotExistsTransformConditionalOperation.cs
@@ -1,7 +1,8 @@
 ﻿using Newtonsoft.Json.Linq;
+using System;
 using System.Collections.Generic;
 
-namespace JSLTSharp.JsonTransforms.Transformations
+namespace JSLTSharp.JsonTransforms.EmbededFunctions
 {
     public class NotExistsTransformConditionalOperation : ExistsConditionalKeyOperation
     {
@@ -11,7 +12,10 @@ namespace JSLTSharp.JsonTransforms.Transformations
         /// <inheritdoc />
         public override bool Apply(JToken dataSource, JToken token, IList<string> parameters)
         {
-            return !Exists(dataSource, token, parameters);
+            if (parameters.Count != 1)
+                throw new InvalidOperationException($"You must provide one select as parameter for function {OperationName}");
+
+            return !Exists(dataSource, parameters[0]);
         }
     }
 }
