+++ src/JSLTSharp/JsonTransforms/EmbededFunctions/ValueTransformations/ToIntegerTransformOperation.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace JSLTSharp.JsonTransforms.Transformations
+namespace JSLTSharp.JsonTransforms.EmbededFunctions
 {
     public class ToIntegerTransformOperation : ExistsTransformOperation
     {
