--- src/JSLTSharp/JsonTransforms/Transformations/ConditionalOperations/ExistsTransformOperation.cs
@@ -1,28 +1,0 @@
-﻿using Newtonsoft.Json.Linq;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-
-namespace JSLTSharp.JsonTransforms.Transformations
-{
-    public class ExistsConditionalKeyOperation : IJsonTransformConditionalCustomOperation
-    {
-        /// <inheritdoc />
-        public virtual string OperationName => "exists";
-
-        /// <inheritdoc />
-        public virtual bool Apply(JToken dataSource, JToken token, IList<string> parameters)
-        {
-            if (parameters.Count != 1)
-                throw new InvalidOperationException($"You must provide only one parameter for function {OperationName}");
-            
-            return Exists(dataSource, token, parameters);
-        }
-
-        protected bool Exists(JToken dataSource, JToken token, IList<string> parameters)
-        {
-            var existsTokenSearch = dataSource.SelectTokens(parameters[0]);
-            return existsTokenSearch != null && existsTokenSearch.Any();
-        }
-    }
-}
